requirements.txt
simplejson
tiktoken
numpy
openai
python-dotenv
gradio